+++ client/src/components/article.tsx
@@ -1,24 +1,32 @@
 'use client';
+
 import { routes } from '@/shared/constants';
-import { colorBorder, colorTextSecondary, secondaryFont, secondaryLink } from '@/theme';
+import { Nullable } from '@/shared/types';
+import { colorTextSecondary, secondaryFont, secondaryLink } from '@/theme';
+import { config } from '@/utils';
 import { EyeFilled } from '@ant-design/icons';
 import { Button, Card, Divider, Space, Typography, Grid, Row, Col } from 'antd';
 import Link from 'next/link';
-import React from 'react';
+import React, { useMemo } from 'react';
 
 const { Title, Text } = Typography;
 const { useBreakpoint } = Grid;
 
 interface ArticleProps {
+  id: number;
+  categoryId: number;
   title: string;
   image: string;
   description: string;
   category: string;
-  views: string;
+  views: Nullable<number>;
 }
 
-export const Article: React.FC<ArticleProps> = ({ title, image, description, category, views }) => {
+export const Article: React.FC<ArticleProps> = ({ title, image, description, category, views, id, categoryId }) => {
   const screens = useBreakpoint();
+
+  const articleUrl = useMemo(() => routes.article.path.replace(':id', id.toString()), [id]);
+  const categoryUrl = useMemo(() => routes.category.path.replace(':id', categoryId.toString()), [categoryId]);
 
   return (
     <Card
@@ -33,20 +41,20 @@ export const Article: React.FC<ArticleProps> = ({ title, image, description, cat
             lineHeight: '35px',
           }}
         >
-          <Link href={routes.article.path.replace(':id', '1')}>{title}</Link>
+          <Link href={articleUrl}>{title}</Link>
         </Title>
       }
     >
       <div style={{ width: '100%', padding: screens.md ? '0 48px' : 0 }}>
         <Link
-          href={routes.article.path.replace(':id', '1')}
+          href={articleUrl}
           style={{
             margin: screens.md ? '48px 0' : '0 0 24px 0',
             display: 'inline-block',
             width: '100%',
             position: 'relative',
             height: screens.md ? 345 : 152,
-            backgroundImage: `url(${image})`,
+            backgroundImage: `url(${config.BPC_URL + image})`,
             backgroundSize: 'cover',
             backgroundPosition: 'center',
           }}
@@ -76,17 +84,17 @@ export const Article: React.FC<ArticleProps> = ({ title, image, description, cat
                 height: '100%',
               }}
             >
-              <Link href={routes.category.path.replace(':id', '1')} style={secondaryLink}>
+              <Link href={categoryUrl} style={secondaryLink}>
                 {category}
               </Link>
               <div>
                 <EyeFilled style={{ marginRight: 8 }} />
-                <span style={{ textTransform: 'uppercase', letterSpacing: '0.25em' }}>{views}</span>
+                <span style={{ textTransform: 'uppercase', letterSpacing: '0.25em' }}>{views || 0}</span>
               </div>
             </Space>
           </Col>
           <Col md={{ span: 12, order: 1 }} sm={24} xs={24}>
-            <Link href={routes.category.path.replace(':id', '1')}>
+            <Link href={articleUrl}>
               <Button key="read-more" size="large" block={!screens.md}>
                 Читати далі
               </Button>