+++ client/src/components/index.ts
@@ -4,3 +4,4 @@ export * from './header';
 export * from './pagination';
 export * from './ads';
 export * from './footer';
+export * from './navbar/navbar';