+++ client/src/shared/types/tag.ts
@@ -1,0 +1,9 @@
+export interface Tag {
+  id: number;
+  attributes: {
+    title: string;
+    createdAt: string;
+    updatedAt: string;
+    publishedAt: string;
+  };
+}